Docker Deploy: Create basic conf
We'd like to get our app deployed on a live site and Docker is the
cool thing to use.

This should be all the conf files needed to deploy our super basic
hello world app to the Google Cloud!

This is mostly following https://www.codemag.com/Article/2111071/Beginner%E2%80%99s-Guide-to-Deploying-PHP-Laravel-on-the-Google-Cloud-Platform