GCloud Deploy: Fix permissions
We're getting further in the deploy process and now it's complaining about
bootstrap/cache directory permissions.

Seems like we're trying to setup things in the composer script. There's
another artisan call that adds caching stuff so we'll try adding that.

It is failing earlier so we'll just try creating the directory first.